Refactor NLS code in separate folder
