+++ examples/llm_compression/torch/qat_with_nls_downstream/pics/lora_vs_nls.png
