+++ examples/llm_compression/torch/qat_with_nls_downstream/requirements.txt
@@ -5,3 +5,4 @@ numpy>=1.23.5,<2
 openvino==2025.1
 optimum-intel>=1.22.0
 transformers>=4.48.0
+lm_eval==0.4.8
