+++ tests/cross_fw/examples/example_scope.json
@@ -288,7 +288,7 @@
     "llm_compression_qat_with_nls": {
         "backend": "torch",
         "device": "cuda",
-        "requirements": "examples/llm_compression/torch/qat_with_lora/requirements.txt",
+        "requirements": "examples/llm_compression/torch/qat_with_nls_downstream/requirements.txt",
         "cpu": "Intel(R) Core(TM) i9-10980XE CPU @ 3.00GHz",
         "accuracy_tolerance": 0.02,
         "accuracy_metrics": {
